feat(shared): extract framework-free core into a workspace package
Groundwork for the Expo mobile app: the record model, validation gate and
controlled vocabularies now live in @cioos/shared, so the web SPA and the
mobile app can share one source of truth rather than growing a second copy.
(server/src/lib/blankRecord.js is already a hand-maintained twin of the
frontend's; a third would be untenable.)

Moved ~3,000 lines across 18 files that had no React, MUI or DOM dependency.
The SPA is untouched apart from thin re-export shims left at the old paths, so
none of its 159 source files needed changing.

Two substantive fixes, both made deliberately and pinned by tests:

- blankRecord now declares the 11 fields the tabs wrote but it never declared
  (resourceType, metadataScope, metadataScopeIso, projects, taxa, noTaxa,
  noVerticalExtent, verticalExtentEPSG), and limitations is {en,fr} to match
  every consumer instead of "".

- Three validators are length-checked rather than bare-truthy, because
  blankRecord is spread over every record on save and [] is truthy. Without
  this, declaring taxa/resourceType would have made an empty record pass its
  own submit gate, and declaring resourceType would have made the spatial
  section required on a brand-new record.

Behaviour is provably unchanged: a blank record still fails exactly the same
15 validators and still scores 6/21, measured before and after the move.

Also exports `validators`, which was previously module-private — the mobile
completeness ledger groups sections by validators[field].tab.

The URL-reachability warning is split into validateWarnings.js and takes
checkURLActive by injection, keeping the validation core network-free so it
runs offline unchanged.
