+++ package.json
@@ -6,7 +6,11 @@
     "node": ">=22.0.0",
     "npm": ">=10.0.0"
   },
+  "workspaces": [
+    "shared"
+  ],
   "dependencies": {
+    "@cioos/shared": "*",
     "@citation-js/date": "^0.5.1",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/modifiers": "^9.0.0",
