+++ shared/src/eovs.js
@@ -1,0 +1,5 @@
+// Synced from cioos-siooc/cioos-commons via .github/sync.yml — do not hand-edit.
+import eovs from "./data/eovs.json";
+import eovCategories from "./data/eovCategories.json";
+
+export { eovs, eovCategories };
