+++ shared/src/platforms.json
