+++ shared/src/themes.js
@@ -1,0 +1,7 @@
+const themesList = [
+  {
+    en: ["Oceanographic", "Biological", "Other"],
+    fr: ["Océanographie", "Biologie", "Autre"],
+  },
+];
+export default themesList;
