+++ src/utils/themes.js
@@ -1,7 +1,3 @@
-const themesList = [
-  {
-    en: ["Oceanographic", "Biological", "Other"],
-    fr: ["Océanographie", "Biologie", "Autre"],
-  },
-];
-export default themesList;
+// Moved to the @cioos/shared workspace so the web SPA and the Expo app share
+// one source of truth. Re-exported here so existing imports keep working.
+export { default } from "@cioos/shared/themes.js";
